backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
flask==3.0.0
python-multipart==0.0.6
requests==2.31.0
numpy==1.24.3
Pillow==10.1.0
joblib==1.3.2
scikit-learn==1.3.2
xgboost==2.0.2
tensorflow-cpu==2.15.0
keras==2.15.0
